Fixed some compilation errors

+++ HLTriggerOffline/Tau/src/HLTTauMcInfo.cc
@@ -111,7 +111,7 @@ void HLTTauMcInfo::produce(edm::Event& iEvent, const edm::EventSetup& iES)
       Handle<PFTauDiscriminatorByIsolation> thePFTauDiscriminatorByIsolation;
       iEvent.getByLabel(pfTauDiscriminatorProd_,thePFTauDiscriminatorByIsolation);
 
-      for(int it =0; it<product_Jets_tmp.size();it++)
+      for(size_t it =0; it<product_Jets_tmp.size();it++)
 	{ 
 	  for (PFTauCollection::size_type iPFTau=0;iPFTau<thePFTauHandle->size();iPFTau++) {
 	    PFTauRef thePFTau(thePFTauHandle,iPFTau);
@@ -125,7 +125,7 @@ void HLTTauMcInfo::produce(edm::Event& iEvent, const edm::EventSetup& iES)
 	}
       
     }else{
-      for(int it =0; it<product_Jets_tmp.size();it++)
+      for(size_t it =0; it<product_Jets_tmp.size();it++)
 	{ 
 	  product_Jets->push_back(product_Jets_tmp[it]);
 	}
